Create IN - OUT Ports
Port A IN. Port B OUT. 
Process 1:
When switches Δ6 & Δ7 are on (the rest are off), all LEDs are ON. Check port A again.

Process 2:
When switches Δ1 & Δ2 & Δ7 are on (the rest are off), turn on LEDs L4,L6 & L7. Three second delay. Then turn on the correct LEDs to show the contents of address 2400H. Three second delay. Check port A again.

Process 3:
When switches Δ4 & Δ6 are on (the rest are off), add FAH to 2FH and store the result in address 2410H. Then perform the XOR operation between the contents of address 2410H and the number A9H. Store the result in address 2415H. Print the result in the data field of the Display of the Mikrokit. Check port A again.